Fix time unit definition (#36)
* Fixes the time unit transformations
* Adds input to be able to decide if all previous comments with the same task lists should all be delete, and a new one created

+++ index.js
@@ -26,7 +26,9 @@ async function run() {
 
           if(completedTasksArr.length == count && count != 0) {
               console.log(`All ${count} tasks have been successfully completed!`);
-              await deleteGithubComment(commentID);
+              if(inputs.deleteCommentAfterCompletion) {
+                await deleteGithubComment(commentID);
+              }
           } else {
               core.setFailed(`Not all tasks have been completed, only ${completedTasksArr.length} out of ${count} have been completed.\n Re-run this job once the task list has been completed.`);
           }
